+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.2 (UnReleased)
+
+FEATURES:
+
+* **New Resource:** `agile_logical_port`
+
+BUG FIXES:
+* Fix tenant always deploying without any change
+
 ## 0.0.1 (December 10, 2021)
 
 - Initial Release